+++ tests/func/test_get.py
@@ -4,8 +4,8 @@ import os
 import pytest
 
 from dvc.cli import main
-from dvc.data.db import ODBManager
 from dvc.fs import system
+from dvc.odbmgr import ODBManager
 from dvc.repo import Repo
 from dvc.repo.get import GetDVCFileError
 from dvc.utils.fs import makedirs
